\name{believe14-package}
\alias{believe14-package}
\alias{believe14}
\docType{package}
\title{\packageTitle{believe14}}
\description{\packageDescription{believe14}}
\section{Package Content}{\packageIndices{believe14}}
\author{\packageAuthor{believe14}}
\section{Maintainer}{\packageMaintainer{believe14}}
\keyword{package}
